Add a note about installing Ruby and Node/npm
When following the install instructions I did not have Ruby installed.
It took me a little while to figure out the proper way to install gems.
MacOS has a global Ruby install that you should not be modifying.
Instead your should use something like `rbenv` to set up a dev install.

This commit adds a link  to an explanation of how to do so.
For completeness sake, it also adds instructions for installing npm.