Replace broken image with Data URI
The footer was hotlinking an image from the Cloud Four site. That image had been moved to a CDN and was no longer at the original location. It seemed like this image should live in this repo to avoid similar issues in the future, but I couldn't figure out the best place to put the image in the repo. The `images` directory seems like it's meant to house PWA icons.

As  a quick solution I inserted this image  as an SVG Data URI